Move datalad files functions to datalad/files.js
