+++ packages/openneuro-server/handlers/datalad.js
@@ -4,7 +4,7 @@ import mime from 'mime-types'
 import { generateDataladCookie } from '../libs/authentication/jwt'
 import { getDraftFiles, getDatasetRevision } from '../datalad/draft'
 import { getSnapshot } from '../datalad/snapshots'
-import { encodeFilePath, decodeFilePath } from '../datalad/dataset.js'
+import { encodeFilePath, decodeFilePath } from '../datalad/files.js'
 
 /**
  * Handlers for datalad dataset manipulation
